requirements.txt
requests
python-dotenv
pytelegrambotapi
pycountry
Babel
